Fix an issue with incorrect spill locations.
MIPS branches define $at as an implicit definition so that the long
branch pass has it available when expanding them to long branches.  The
CHERI branches did not have this.  This meant that $at was spilled in
between the conditional branch and the (dead, fall-through, will-be-
removed-later) unconditional branch at the end of a basic block.  This,
in turn, meant that the fast register allocator's check to find the
first terminator didn't find the conditional branch, because it
performed a linear scan backwards to find the first instruction after
the last non-terminator.  This then resulted in all of the additional
spills being after the conditional branch.

Fixes #254